+++ demo/blocks_scroll/run.py
@@ -1,6 +1,5 @@
 import gradio as gr
 
-
 demo = gr.Blocks()
 
 with demo:
@@ -11,13 +10,13 @@ with demo:
     <div style='height: 800px; width: 100px; background-color: pink;'></div>
     """)
     out = gr.Textbox()
-    
-    scroll_btn.click(lambda x: x, 
-               inputs=inp, 
+
+    scroll_btn.click(lambda x: x,
+               inputs=inp,
                outputs=out,
                 scroll_to_output=True)
-    no_scroll_btn.click(lambda x: x, 
-               inputs=inp, 
+    no_scroll_btn.click(lambda x: x,
+               inputs=inp,
                outputs=out)
 
 if __name__ == "__main__":