+++ demo/zip_files/run.py
@@ -3,7 +3,6 @@ from zipfile import ZipFile
 
 import gradio as gr
 
-
 def zip_files(files):
     with ZipFile("tmp.zip", "w") as zipObj:
         for idx, file in enumerate(files):
@@ -14,9 +13,9 @@ demo = gr.Interface(
     zip_files,
     gr.File(file_count="multiple", file_types=["text", ".json", ".csv"]),
     "file",
-    examples=[[[os.path.join(os.path.dirname(__file__),"files/titanic.csv"), 
-    os.path.join(os.path.dirname(__file__),"files/titanic.csv"), 
-    os.path.join(os.path.dirname(__file__),"files/titanic.csv")]]], 
+    examples=[[[os.path.join(os.path.dirname(__file__),"files/titanic.csv"),
+    os.path.join(os.path.dirname(__file__),"files/titanic.csv"),
+    os.path.join(os.path.dirname(__file__),"files/titanic.csv")]]],
     cache_examples=True
 )