+++ gradio/helpers.py
@@ -415,7 +415,10 @@ class Examples:
             show_api=False,
         )
 
-    async def async_lazy_cache(self, example_index, *input_values):
+    async def async_lazy_cache(
+        self, example_value: tuple[int, list[Any]], *input_values
+    ):
+        example_index, _ = example_value
         cached_index = self._get_cached_index_if_cached(example_index)
         if cached_index is not None:
             output = self.load_from_cache(cached_index)
@@ -433,7 +436,8 @@ class Examples:
         with open(self.cached_indices_file, "a") as f:
             f.write(f"{example_index}\n")
 
-    def sync_lazy_cache(self, example_index, *input_values):
+    def sync_lazy_cache(self, example_value: tuple[int, list[Any]], *input_values):
+        example_index, _ = example_value
         cached_index = self._get_cached_index_if_cached(example_index)
         if cached_index is not None:
             output = self.load_from_cache(cached_index)
@@ -517,7 +521,7 @@ class Examples:
                     )
                 output = prediction["data"]
                 if len(generated_values):
-                    output = merge_generated_values_into_output(
+                    output = await merge_generated_values_into_output(
                         self.outputs, generated_values, output
                     )
                 if self.batch:
@@ -579,7 +583,7 @@ class Examples:
         return output
 
 
-def merge_generated_values_into_output(
+async def merge_generated_values_into_output(
     components: Sequence[Component], generated_values: list, output: list
 ):
     from gradio.components.base import StreamingOutput
@@ -594,9 +598,11 @@ def merge_generated_values_into_output(
                 if isinstance(processed_chunk, (GradioModel, GradioRootModel)):
                     processed_chunk = processed_chunk.model_dump()
                 binary_chunks.append(
-                    output_component.stream_output(processed_chunk, "", i == 0)[0]
+                    (await output_component.stream_output(processed_chunk, "", i == 0))[
+                        0
+                    ]
                 )
-            binary_data = b"".join(binary_chunks)
+            binary_data = b"".join([d["data"] for d in binary_chunks])
             tempdir = os.environ.get("GRADIO_TEMP_DIR") or str(
                 Path(tempfile.gettempdir()) / "gradio"
             )