+++ js/audio/player/AudioPlayer.svelte
@@ -11,6 +11,8 @@
 	import type { WaveformOptions } from "../shared/types";
 	import { createEventDispatcher } from "svelte";
 
+	import Hls from "hls.js";
+
 	export let value: null | FileData = null;
 	$: url = value?.url;
 	export let label: string;
@@ -40,6 +42,9 @@
 
 	let show_volume_slider = false;
 	let audio_player: HTMLAudioElement;
+
+	let audio_player: HTMLAudioElement;
+	let stream_active = false;
 
 	const dispatch = createEventDispatcher<{
 		stop: undefined;
@@ -130,6 +135,7 @@
 	};
 
 	async function load_audio(data: string): Promise<void> {
+		stream_active = false;
 		await resolve_wasm_src(data).then((resolved_src) => {
 			if (!resolved_src || value?.is_stream) return;
 			return waveform?.load(resolved_src);
@@ -137,6 +143,52 @@
 	}
 
 	$: url && load_audio(url);
+
+	function load_stream(value: FileData | null): void {
+		if (!value || !value.is_stream || !value.url) return;
+		if (!audio_player) return;
+		if (Hls.isSupported() && !stream_active) {
+			// Set config to start playback after 1 second of data received
+			const hls = new Hls({
+				maxBufferLength: 1,
+				maxMaxBufferLength: 1,
+				lowLatencyMode: true
+			});
+			hls.loadSource(value.url);
+			hls.attachMedia(audio_player);
+			hls.on(Hls.Events.MANIFEST_PARSED, function () {
+				if (waveform_settings.autoplay) audio_player.play();
+			});
+			hls.on(Hls.Events.ERROR, function (event, data) {
+				console.error("HLS error:", event, data);
+				if (data.fatal) {
+					switch (data.type) {
+						case Hls.ErrorTypes.NETWORK_ERROR:
+							console.error(
+								"Fatal network error encountered, trying to recover"
+							);
+							hls.startLoad();
+							break;
+						case Hls.ErrorTypes.MEDIA_ERROR:
+							console.error("Fatal media error encountered, trying to recover");
+							hls.recoverMediaError();
+							break;
+						default:
+							console.error("Fatal error, cannot recover");
+							hls.destroy();
+							break;
+					}
+				}
+			});
+			stream_active = true;
+		} else if (!stream_active) {
+			audio_player.src = value.url;
+			if (waveform_settings.autoplay) audio_player.play();
+			stream_active = true;
+		}
+	}
+
+	$: load_stream(value);
 
 	onMount(() => {
 		window.addEventListener("keydown", (e) => {
@@ -152,7 +204,7 @@
 
 <audio
 	class="standard-player"
-	class:hidden={!value || !value.is_stream}
+	class:hidden={!(value && value.is_stream)}
 	controls
 	autoplay={waveform_settings.autoplay}
 	on:load