+++ js/video/shared/InteractiveVideo.svelte
@@ -120,6 +120,7 @@
 					{autoplay}
 					src={value.url}
 					subtitle={subtitle?.url}
+					is_stream={false}
 					on:play
 					on:pause
 					on:stop